[charts] DRY post schemas
